+++ CHANGELOG.md
@@ -1,4 +1,10 @@
+
 ## Release History
+* 0.3.4
+  - add backend factory
+  - add context creating functions
+  - add probe function to ess_backend as virtual
+  - add esp32 i2s_generic_backend
 * 0.3.3
   - rename project to OpenESS ++
   - switch to C++